refactor evaluator results
