+++ ra-optimisation/search-tree-rao/src/main/java/com/powsybl/openrao/searchtreerao/commons/costevaluatorresult/MaxCostEvaluatorResult.java
@@ -17,13 +17,13 @@ import java.util.stream.DoubleStream;
 /**
  * @author Thomas Bouquet {@literal <thomas.bouquet at rte-france.com>}
  */
-public class MaxCostEvaluatorResult extends AbstractCostEvaluatorResult {
+public class MaxCostEvaluatorResult extends AbstractStateWiseCostEvaluatorResult {
     public MaxCostEvaluatorResult(Map<State, Double> costPerState, List<FlowCnec> costlyElements) {
-        super(costPerState, costlyElements, -Double.MAX_VALUE);
+        super(costPerState, costlyElements);
     }
 
     @Override
-    protected double evaluateResultsWithSpecificStrategy(double preventiveCost, DoubleStream postContingencyCosts) {
-        return Math.max(preventiveCost, postContingencyCosts.max().orElse(defaultCost));
+    protected double evaluateResultsWithSpecificStrategy(DoubleStream filteredCostsStream) {
+        return filteredCostsStream.max().orElse(-Double.MAX_VALUE);
     }
 }
