+++ ra-optimisation/search-tree-rao/src/main/java/com/powsybl/openrao/searchtreerao/commons/objectivefunctionevaluator/SensitivityFailureOvercostEvaluator.java
@@ -11,7 +11,7 @@ import com.powsybl.openrao.data.cracapi.State;
 import com.powsybl.openrao.data.cracapi.cnec.Cnec;
 import com.powsybl.openrao.data.cracapi.cnec.FlowCnec;
 import com.powsybl.openrao.data.raoresultapi.ComputationStatus;
-import com.powsybl.openrao.searchtreerao.commons.costevaluatorresult.ConstantCostEvaluatorResult;
+import com.powsybl.openrao.searchtreerao.commons.costevaluatorresult.AbsoluteCostEvaluatorResult;
 import com.powsybl.openrao.searchtreerao.commons.costevaluatorresult.CostEvaluatorResult;
 import com.powsybl.openrao.searchtreerao.result.api.FlowResult;
 import com.powsybl.openrao.searchtreerao.result.api.RemedialActionActivationResult;
@@ -43,14 +43,14 @@ public class SensitivityFailureOvercostEvaluator implements CostEvaluator {
     public CostEvaluatorResult evaluate(FlowResult flowResult, RemedialActionActivationResult remedialActionActivationResult) {
         if (flowResult.getComputationStatus() == ComputationStatus.FAILURE) {
             TECHNICAL_LOGS.info(String.format("Sensitivity failure : assigning virtual overcost of %s", sensitivityFailureOvercost));
-            return new ConstantCostEvaluatorResult(sensitivityFailureOvercost);
+            return new AbsoluteCostEvaluatorResult(sensitivityFailureOvercost);
         }
         for (State state : states) {
             if (flowResult.getComputationStatus(state) == ComputationStatus.FAILURE) {
                 TECHNICAL_LOGS.info(String.format("Sensitivity failure for state %s : assigning virtual overcost of %s", state.getId(), sensitivityFailureOvercost));
-                return new ConstantCostEvaluatorResult(sensitivityFailureOvercost);
+                return new AbsoluteCostEvaluatorResult(sensitivityFailureOvercost);
             }
         }
-        return new ConstantCostEvaluatorResult(0.0);
+        return new AbsoluteCostEvaluatorResult(0.0);
     }
 }
